backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.0
python-multipart>=0.0.9
aiofiles>=23.2.1
pillow>=10.0.0
python-dotenv>=1.0.0
# AI providers
google-cloud-aiplatform>=1.60.0
google-generativeai>=0.8.0
# URL fetching + HTML parsing for recipe import
httpx>=0.27.0
beautifulsoup4>=4.12.0
